+++ src/app/Components/ShowItem.tsx
@@ -46,6 +46,7 @@ const ShowItem: React.FC<ShowItemProps> = ({
               Delete
             </button>
             <Toaster />
+            
           </div>
         ))}
       </div>
