+++ src/app/layout.tsx
@@ -1,26 +1,16 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
-import { Toaster } from "react-hot-toast";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
 
 export default function RootLayout({
   children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+}: {
+  children: React.ReactNode
+}) {
   return (
     <html lang="en">
-      <body className={inter.className}>
-        <Toaster position="top-center" />
-        {children}
-      </body>
+      <body>{children}</body>
     </html>
-  );
+  )
 }
